saint_os/firmware/rpi5/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saint-node"
version = "1.0.0"
description = "SAINT.OS Raspberry Pi 5 Node Firmware"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SAINT.OS Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
saint-node = "saint_node.node:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["saint_node*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
